Move rfactor python version
